backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
psycopg[binary]>=3.2.0
psycopg-pool>=3.2.0
bcrypt>=4.0.0
PyJWT>=2.0.0
python-multipart>=0.0.9
